requirements.txt
flask==2.0.2
flask-pymongo==2.3.0
pandas==1.4.0
python-dotenv==0.19.2